WIP: Replace RelationSource with Analyzed-/QueriedRelation
In order to do a "query push-down" we used a custom data structure to
represent a query with a JOIN.

For example

    select t1.x, t2.x from t1, t2 where t1.x = 10

Was represented as:

    select t1.x, t2.x from t1, t2

    t1 querySpec:
        outputs: [x]
        where: x = 10

    t2 querySpec:
        outputs: [x]

This commit changes the internal representation so that the sub-query
data structures are used. The above example now looks like this:

    select t1.x, t2.x from
        (select x from t1 where x = 1) t1,
        (select x from t2) t2

This should make it easier to support joined virtual tables since the
internal representation is already the same.